gm_18A5 matches

+++ src/melee/mn/mnstagesel.c
@@ -529,7 +529,7 @@ void mnStageSel_8025BB5C_OnLeave(UNK_T unused)
     }
 }
 
-u8 mnStageSel_8025BBD4(void)
+int mnStageSel_8025BBD4(void)
 {
     return mnStageSel_803F06D0[mnStageSel_802599EC()].xB;
 }
